pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-search-intent-first"
version = "0.1.0"
description = "Intent-first photo search app (Streamlit + CLIP)"
authors = [{name = "Your Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "streamlit>=1.32.0",
  "sentence-transformers>=2.7.0",
  "torch>=2.1.0",
  "Pillow>=10.0.0",
  "numpy>=1.24.0",
  "tqdm>=4.66.0",
  "requests>=2.31.0",
  "openai>=1.52.0",
]

[project.optional-dependencies]
ann = ["annoy>=1.17.3"]
faiss = ["faiss-cpu>=1.7.4"]
ocr = ["easyocr>=1.7.1"]
hnsw = ["hnswlib>=0.7.0"]
dev = ["ruff>=0.5.0", "mypy>=1.8.0", "pytest>=7.4.0"]

[project.scripts]
ps-intent = "photo_search_intent_first.cli:main"
ps-intent-ui = "photo_search_intent_first.launcher:main"

[tool.setuptools]
package-dir = {"" = "photo-search-intent-first"}

[tool.setuptools.packages.find]
where = ["photo-search-intent-first"]

[tool.ruff]
line-length = 100
target-version = "py39"
